refactor(window): split ProjectViewManager into sibling modules
- Extract paint-gate bridging into ProjectViewPaintGateController.ts
- Extract view creation/load into ProjectViewFactory.ts
- Extract activation/deactivation lifecycle into ProjectViewLifecycleController.ts
- Extract LRU/memory-pressure eviction into ProjectViewEvictionController.ts
- Extract agent-state cache into ProjectViewAgentStateCache.ts
- Extract WebContents handlers + crash recovery into ProjectViewHandlers.ts
- Extract performSwitch into ProjectViewSwitchController.ts
- Move shared internal types into ProjectViewManagerTypes.ts
- ProjectViewManager.ts remains the sole exported class with an unchanged
  public API; extracted modules operate on its state via an explicit
  host parameter
